+++ Debian/setup_debian.sh
@@ -153,14 +153,6 @@ prepare_env_etc
 msg_1 "apt upgrade"
 apt upgrade -y
 
-if [ -n "$DEB_PKGS" ]; then
-    msg_1 "Add Debian packages"
-    echo "$DEB_PKGS"
-    #  shellcheck disable=SC2086
-    apt install -y $DEB_PKGS
-fi
-echo
-
 if [ -n "$DEB_PKGS_SKIP" ]; then
     msg_1 "Removing Debian packages"
     echo "$DEB_PKGS_SKIP"
@@ -171,6 +163,14 @@ if [ -n "$DEB_PKGS_SKIP" ]; then
     #
     #  shellcheck disable=SC2086
     apt purge -y $DEB_PKGS_SKIP
+fi
+echo
+
+if [ -n "$DEB_PKGS" ]; then
+    msg_1 "Add Debian packages"
+    echo "$DEB_PKGS"
+    #  shellcheck disable=SC2086
+    apt install -y $DEB_PKGS
 fi
 echo