+++ Alpine/setup_alpine.sh
@@ -28,11 +28,13 @@ install_apks() {
 
     #
     #  Install some custom apks, where the current repo version cant
-    #  be used
+    #  be used, so we use the last known to work on Debian version
+    #  Since they are installed as a file, they are pinned, and wont
+    #  be replaced by an apt upgrade
     #
     msg_2 "Custom apks"
     if wget https://dl-cdn.alpinelinux.org/alpine/v3.10/main/x86/mtr-0.92-r0.apk >/dev/null 2>&1 ; then
-	msg_3 "mtr is a full screen traceroute"
+	msg_3 "mtr - a full screen traceroute"
 	apk add ./mtr-0.92-r0.apk && rm mtr-0.92-r0.apk
     fi
 }