+++ common_AOK/setup_final_tasks.sh
@@ -149,23 +149,6 @@ user_interactions
 
 hostfs_is_alpine && install_aok_apks
 
-#
-#  Removing default host
-#
-hostn_service=/etc/init.d/hostname
-if [ -f "$hostn_service" ]; then
-    msg_2 "Disabling hostname service not working on iSH"
-    mv -f "$hostn_service" /etc/init.d/NOT-hostname
-fi
-
-if hostfs_is_debian || hostfs_is_devuan; then
-    msg_3 "Removing hostname service files not meaningfull on iSH"
-    rm -f /etc/init.d/hostname
-    rm -f /etc/init.d/hostname.sh
-    rm -f /etc/rcS.d/S01hostname.sh
-    rm -f /etc/systemd/system/hostname.service
-fi
-
 "$aok_content"/common_AOK/aok_hostname/set_aok_hostname.sh
 
 set_initial_login_mode