+++ CHANGES.md
@@ -4,6 +4,7 @@ I will try to keep track of changes between releases here
 
 ## Changes newest first
 
+- Added my package adm tools Mapt & Mapk
 - Improved detection if already chrooted
 - Removed DEBUG_BUILD havent used it in a while and that stuff was going obsolete
 - Added check that sudo is installed